Moved the css into its owns specfic folder and formatted the navbar.jsx code
