+++ package.json
@@ -19,11 +19,16 @@
     "@babel/preset-react": "^7.27.1",
     "axios": "^1.10.0",
     "babel-loader": "^10.0.0",
+    "class-variance-authority": "^0.7.1",
+    "clsx": "^2.1.1",
     "dotenv": "^17.1.0",
+    "lucide-react": "^0.536.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "react-router-dom": "^7.6.3",
     "react-youtube": "^10.1.0",
+    "tailwind-merge": "^3.3.1",
+    "tw-animate-css": "^1.3.6",
     "webpack": "^5.99.9",
     "webpack-cli": "^6.0.1"
   },
