+++ src/components/SocialMedia.jsx
@@ -1,0 +1,10 @@
+import React from "react";
+
+const SocialMedia = () =>{
+    return(
+        <h2>Hello World</h2>
+    )
+
+}
+
+export default SocialMedia;
